% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_record.R
\name{add_empty_row}
\alias{add_empty_row}
\title{Add an Empty Row to a Data Frame}
\usage{
add_empty_row(.data)
}
\arguments{
\item{.data}{A data frame to which an empty row will be added}
}
\value{
Modified data frame with an additional empty row
}
\description{
Adds a single row of NA values to a data frame
}
