% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_dots}
\alias{deparse_dots}
\title{Deparse dots arguments into character vector}
\usage{
deparse_dots(...)
}
\arguments{
\item{...}{Arguments to deparse}
}
\value{
A character vector containing the deparsed expressions
}
\description{
Takes dots arguments and deparses them into a character vector.
}
\keyword{internal}
