% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dots_n}
\alias{dots_n}
\title{Count number of dots arguments}
\usage{
dots_n(...)
}
\arguments{
\item{...}{Arguments to count}
}
\value{
Integer length of dots arguments
}
\description{
Returns the number of arguments passed via ...
}
\keyword{internal}
