% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equal_names}
\alias{equal_names}
\title{Check if two objects have identical column names}
\usage{
equal_names(x, y)
}
\arguments{
\item{x}{First object to compare}

\item{y}{Second object to compare}
}
\value{
Logical indicating if objects have identical column names
}
\description{
Compares column names between two objects element by element.
}
\keyword{internal}
