% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract}
\alias{extract}
\title{Extract elements}
\usage{
extract()
}
\value{
typically an array-like R object of a similar class as x.
}
\description{
Alias for base R extract operator `[`
}
\keyword{internal}
