% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract2}
\alias{extract2}
\title{Extract single element}
\usage{
extract2()
}
\value{
typically an array-like R object of a similar class as x.
}
\description{
Alias for base R extract operator `[[`
}
\keyword{internal}
