% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Check if object is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if object has zero length
}
\description{
Tests if an object has length zero.
}
\keyword{internal}
