% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nested_list}
\alias{is_nested_list}
\title{Determine if list is nested}
\usage{
is_nested_list(x)
}
\arguments{
\item{x}{List to check}
}
\value{
Logical indicating if list is nested
}
\description{
Checks for nested lists in batch functions
}
\keyword{internal}
