% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz-package.R
\docType{package}
\name{matriz-package}
\alias{matriz}
\alias{matriz-package}
\title{matriz: Matrix synthesis tools for epidemiologic and health science research; specifically, literature reviews}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

matriz provides tools to create, update and fill literature matrices commonly used in research, specifically epidemiologic and health sciences research. The project is born out of need as a easy–to–use tool for my research methods classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpmonteagudo28/matriz}
  \item Report bugs at \url{https://github.com/jpmonteagudo28/matriz/issues}
}

}
\author{
\strong{Maintainer}: JP Monteagudo \email{jpmonteagudo2014@gmail.com} (\href{https://orcid.org/0009-0003-6465-6658}{ORCID}) [copyright holder]

}
\keyword{internal}
