% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{matriz_message}
\alias{matriz_message}
\title{Display package version for \pkg{matriz}}
\usage{
matriz_message()
}
\value{
\code{dmatriz_message()} returns a message about the install version
of \pkg{matriz}.
}
\description{
\code{matriz_message()} produces a message about the package version
and the version of R making use of this package.
}
\examples{
matriz_message()


}
\author{
JP Monteagudo
}
