% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_names.R
\name{matriz_names}
\alias{matriz_names}
\title{Retrieve Column Classes from deafult literature matrix.}
\usage{
matriz_names(...)
}
\arguments{
\item{...}{extra arguments to pass as column names for the lierature matrix}
}
\value{
A data frame with one column named \code{class} that lists the class
  of each column from the matrix or data frame returned by \code{init_matrix()}.
}
\description{
This function calls \code{init_matrix()} to obtain a matrix or data frame,
then extracts the class of each column. It returns a data frame containing
the class information for each column.
}
\details{
The purpose of this function is to provide the user with a quick way to check the default names and classes
as the matrix is being filled instead of having to type `str(init_matrix())` every time the user forgets a category in the default matrix.
}
\examples{
\dontrun{
  # Assuming init_matrix() is defined and returns a valid data frame or matrix
  class_info <- matriz_names()
  print(class_info)
}

}
