% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rid_dups}
\alias{rid_dups}
\title{Remove Duplicates from Vectors or Data Frame Columns}
\usage{
rid_dups(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A vector or data frame}

\item{incomparables}{A vector of values that cannot be compared. See ?duplicated}

\item{...}{arguments for particular methods used in 'unique' and 'duplicated'}
}
\value{
The input with duplicates removed
}
\description{
Remove Duplicates from Vectors or Data Frame Columns
}
\keyword{internal}
