% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_column}
\alias{same_column}
\title{Check if two objects have the same number of columns}
\usage{
same_column(x, y)
}
\arguments{
\item{x}{First object to compare}

\item{y}{Second object to compare}
}
\value{
Logical indicating if objects have same number of columns
}
\description{
Compares number of columns between two objects.
}
\keyword{internal}
