% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_length}
\alias{same_length}
\title{Check if two objects have the same length}
\usage{
same_length(x, y)
}
\arguments{
\item{x}{First object to compare}

\item{y}{Second object to compare}
}
\value{
Logical indicating if objects have same length
}
\description{
Compares lengths of two objects, handling special cases for lists and data frames.
}
\keyword{internal}
