% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_char}
\alias{to_char}
\title{Convert Symbol to Character}
\usage{
to_char(symbol)
}
\arguments{
\item{symbol}{A symbol to convert}
}
\value{
A character string
}
\description{
Converts a symbol to a character string
}
\keyword{internal}
